move puff from gameplay into core/spore_cloud
