+++ src/plugins/core/Cargo.toml
@@ -8,3 +8,4 @@ bevy.workspace = true
 bevy_framepace.workspace = true
 avian3d.workspace = true
 blenvy.workspace = true
+bevy_hanabi.workspace = true
